requirements.txt
pandas==2.2.3
numpy==2.2.4
awswrangler[s3,glue,athena]==3.14.0 
boto3==1.35.21
botocore==1.35.21
pyarrow==17.0.0                       
s3fs==2024.6.1
SQLAlchemy==1.4.54
psycopg2-binary==2.9.9
python-dotenv==1.1.1